Fix warning in bget : set bufsize to size_t
Define bufsize in bget to size_t. This avoids warnings due to
comparison of bufsize (which was defined as long) to size of
objects (which would be size_t)

Also refactoring the buffer report function to pass in args with
the correct type (instead of assuming longs pass in args of type
bufsize as needed) and use the appropriate casts (and inits)

+++ src/clib/util/bget.h
@@ -11,7 +11,9 @@
 //#undef NDEBUG
 //#endif
 
-typedef long bufsize;
+#include <cstddef>
+
+typedef size_t bufsize;
 
 #if defined(__cplusplus)
 extern "C" {
